check directory existence

+++ src/NServiceBus.AcceptanceTests/Core/Hosting/When_endpoint_starts.cs
@@ -16,7 +16,10 @@
         {
             // TestContext.CurrentContext.Test.ID is stable across test runs, 
             // therefore we need to clear existing diagnostics file to avoid asserting on a stale file
-            Directory.Delete(basePath, true);
+            if (Directory.Exists(basePath))
+            {
+                Directory.Delete(basePath, true);
+            }
 
             await Scenario.Define<Context>()
                 .WithEndpoint<MyEndpoint>()
